requirements.txt
Django>=4.1.2
djangorestframework
django-cors-headers
Pillow
qrcode
django-filter
requests

